New Rules created: {readEnvRule, readInputRule} Rules separated in different packages
