+++ lexer/src/main/kotlin/rule/basic/TypeIdRule.kt
@@ -1,23 +1,18 @@
-package rule
+package rule.basic
 
-import token.Boolean
+import rule.TokenRule
 import token.Position
 import token.Token
+import token.TokenInterface
 import token.TypeId
 
-class BooleanRule : TokenRule {
-
-    private val keywords = mapOf(
-        "boolean" to TypeId,
-        "true" to Boolean,
-        "false" to Boolean
-    )
-
-    override fun match(input: String, position: Position): Token? {
-        for ((keyword, tokenType) in keywords) {
+class TypeIdRule : TokenRule {
+    override fun match(input: String, position: Position): TokenInterface? {
+        val typeIdKeywords = listOf("String", "Number")
+        for (keyword in typeIdKeywords) {
             if (input.startsWith(keyword)) {
                 return Token(
-                    tokenType, keyword,
+                    TypeId, keyword,
                     Position(
                         position.row,
                         position.startColumn,
