+++ lexer/src/main/kotlin/rule/control/ElseRule.kt
@@ -1,5 +1,6 @@
-package rule
+package rule.control
 
+import rule.TokenRule
 import token.Else
 import token.Position
 import token.Token
