+++ lexer/src/main/kotlin/rule/expression/DeclarationRule.kt
@@ -1,5 +1,6 @@
-package rule
+package rule.expression
 
+import rule.TokenRule
 import token.Declaration
 import token.Position
 import token.Token
