+++ lexer/src/main/kotlin/rule/inherent/PrintlnRule.kt
@@ -1,0 +1,3 @@
+package rule.inherent
+
+class PrintlnRule : BaseNativeMethodRule("println")
