+++ lexer/src/main/kotlin/rule/inherent/ReadEnvRule.kt
@@ -1,0 +1,3 @@
+package rule.inherent
+
+class ReadEnvRule : BaseNativeMethodRule("readEnv")
