+++ lexer/src/main/kotlin/rule/inherent/ReadInputRule.kt
@@ -1,0 +1,3 @@
+package rule.inherent
+
+class ReadInputRule : BaseNativeMethodRule("readInput")
