+++ lexer/src/test/kotlin/textfile/testv2/file5
@@ -1,0 +1,2 @@
+readEnv("DATABASE_URL");
+readInput("message");
